+++ NineInTeam/NineInTeam/Presentation/Membership/DTO/UserUpdateResponse.swift
@@ -1,0 +1,12 @@
+//
+//  UserUpdateResponse.swift
+//  NineInTeam
+//
+//  Created by HeonJin Ha on 2023/06/20.
+//
+
+import Foundation
+
+struct UserUpdateResponse: Codable {
+    let detail: UserUpdateApiModel
+}
