+++ NineInTeam/NineInTeam/Presentation/Membership/SignIn/SignInProviderType.swift
@@ -1,5 +1,5 @@
 //
-//  LoginServiceProvider.swift
+//  SignInProviderType.swift
 //  NineInTeam
 //
 //  Created by HeonJin Ha on 2023/06/18.
@@ -8,6 +8,6 @@
 import Foundation
 
 // apple, none.. 등 로그인 서비스에 따른 분기작업시 사용
-enum SignInServiceProvider: Codable {
+enum SignInProviderType: Codable {
     case kakao
 }
